refactor: 💡 extract code that used in client and addon
