+++ src/utils/viewMode.ts
@@ -1,0 +1,10 @@
+import { ImagesViewMode } from '../types';
+
+export const VIEW_MODE_KEY = 'Creevey_view_mode';
+
+export const viewModes: ImagesViewMode[] = ['side-by-side', 'swap', 'slide', 'blend'];
+
+export const getViewMode = (): ImagesViewMode => {
+  const item = localStorage.getItem(VIEW_MODE_KEY);
+  return item && viewModes.includes(item as ImagesViewMode) ? (item as ImagesViewMode) : 'side-by-side';
+};
